Fix nondeterministic behavior in world mapping.
The current mapping of world string to world from Bukkit was dependent
on world order in Bukkit if multiple worlds had a similar string.

Example, if Bukkit gave the world listing as:
- world_nether
- world_the_end
- world
and the string "world" was trying to be mapped, it would be converted to
"world_nether" instead of the map named "world".

This is now fixed as an exact match takes precedence, and a partial
match will only be accepted if it maps to a single world.